add util methods to obtain an iroh client
